pyproject.toml
[tool.poetry]
name = "floto"
version = "0.0.1"
description = "FLOTO dashboard and API"
authors = ["Chameleon Cloud <contact@chameleoncloud.org>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
djangorestframework = "^3.14.0"
mysqlclient = "^2.1.1"
balena-sdk = "^13.4.0"
django-bootstrap-v5 = "^1.0.11"
paramiko = "^3.0.0"
mozilla-django-oidc = "^3.0.0"
python-keycloak = "^2.13.2"
kubernetes = "^27.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "^6.0.0"
Faker = "^16.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
